+++ round-based/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.3.1
+* Add `rounds_router::simple_store::RoundMsgs::into_iter_including_me()` [#9]
+
+[#9] https://github.com/dfns/round-based/pull/9
+
 ## v0.3.0
 * Add no_std and wasm support [#6]
 * Add state machine wrapper that provides sync API to carry out the protocol defined as async function [#7]